requirements-setup.txt
# =====================================================
# Dependências para Scripts de Setup do Yellow Taxi Dataset
# =====================================================
# 
# Para instalar:
#   pip install -r requirements-setup.txt
#
# Scripts que usam estas dependências:
#   - db_init/setup_yellow_taxi_complete.py
#   - db_init/02_load_yellow_taxi_parquet.py
# =====================================================

# Biblioteca para conectar ao SQL Server via ODBC
pyodbc

# Biblioteca para carregar variáveis de ambiente do .env
python-dotenv==1.2.1

# Bibliotecas para carregamento de dados (download e processamento)
pandas
requests

# Biblioteca para ler arquivos Parquet
pyarrow

# Biblioteca para pandas.to_sql (engine de banco de dados)
sqlalchemy
